typechecking for wishlist

+++ service/routes.py
@@ -166,6 +166,8 @@ class WishlistResource(Resource):
         This endpoint will update a Wishlist based the body that is posted
         """
         app.logger.info('Request to Update a wishlist with id [%s]', wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         check_content_type("application/json")
         wishlist = Wishlist.find(wishlist_id)
         if not wishlist:
@@ -194,6 +196,8 @@ class WishlistResource(Resource):
         This endpoint will delete a Wishlist based the id specified in the path
         """
         app.logger.info('Request to Delete a wishlist with id [%s]', wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find(wishlist_id)
         if wishlist:
             wishlist.delete()
@@ -248,6 +252,8 @@ class WishlistCollection(Resource):
         wishlist = Wishlist()
         app.logger.debug('Payload = %s', api.payload)
         wishlist.deserialize(api.payload)
+        if not Wishlist.correct_type_or_400(wishlist.customer_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Customer ID needs to be a positive integer.")
         wishlist.create()
         app.logger.info('Wishlist with new id [%s] created!', wishlist.id)
         location_url = api.url_for(WishlistResource, wishlist_id=wishlist.id, _external=True)
@@ -354,6 +360,8 @@ class ProductsCollections(Resource):
     def post(self,wishlist_id):
         # ADD A ITEM TO AN WISHLIST
         app.logger.info("Request to add an item to an wishlist")
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         app.logger.debug('Payload = %s', api.payload)
 
@@ -376,6 +384,8 @@ class ProductsCollections(Resource):
     def get(self, wishlist_id):
         """Returns all of items of a wishlist"""
         app.logger.info("Request for Wishlist Products...")
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         results = [product.serialize() for product in wishlist.products]
         return results, status.HTTP_200_OK
@@ -411,6 +421,8 @@ class ProductsResource(Resource):
         """
         app.logger.info(
             "Request to get an item with id: %s from wishlist with id: %s", item_id, wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         product = Product.find_or_404(item_id)
         return product.serialize(), status.HTTP_200_OK
@@ -423,6 +435,8 @@ class ProductsResource(Resource):
         """
         app.logger.info(
             "Request to delete product with id: %s from wishlist with id: %s", item_id, wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         product = Product.find(item_id)
         if product:
             product.delete()
